refactor(compile time): optimize includes, split hpack
