+++ include/cocaine/context.hpp
@@ -98,12 +98,6 @@ private:
 
 };
 
-//template<class Category, class... Args>
-//typename api::category_traits<Category>::ptr_type
-//context_t::get(const std::string& type, Args&&... args) const {
-//    return m_repository->get<Category>(type, std::forward<Args>(args)...);
-//}
-
 } // namespace cocaine
 
 #endif
