+++ include/cocaine/hpack/header_definitions.hpp
@@ -20,6 +20,9 @@
 
 #pragma once
 
+namespace cocaine {
+namespace hpack {
+
 struct headers {
     friend struct header_static_table_t;
 
@@ -865,3 +868,6 @@ public:
         }
     };
 };
+
+} //  namespace hpack
+} //  namespace cocaine
