+++ include/cocaine/trace/trace.hpp
@@ -23,8 +23,9 @@
 
 #include "cocaine/common.hpp"
 
-#include <boost/optional.hpp>
+#include <boost/optional/optional_fwd.hpp>
 
+#include <functional>
 #include <string>
 
 namespace cocaine {
