+++ src/actor_unix.cpp
@@ -1,17 +1,16 @@
 #include "cocaine/rpc/actor_unix.hpp"
-#include "cocaine/rpc/dispatch.hpp"
-
-#include <boost/filesystem/operations.hpp>
 
 #include "cocaine/context.hpp"
-#include "cocaine/logging.hpp"
-
 #include "cocaine/detail/chamber.hpp"
 #include "cocaine/engine.hpp"
+#include "cocaine/logging.hpp"
+#include "cocaine/memory.hpp"
+#include "cocaine/rpc/dispatch.hpp"
 
 #include <blackhole/attribute.hpp>
 #include <blackhole/logger.hpp>
 
+#include <boost/filesystem/operations.hpp>
 #include <boost/lexical_cast.hpp>
 
 using namespace cocaine;
